Refactor : api분리 -> shared > apis > apis.ts 파일에는 공통으로 사용되는 apis만 있고 나머지는 각 폴더의 service에서 관리하도록 함
